+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### feat: support dogecoin for the local dev environment
 
-You can now launch a network with `dfx network start --enable-dogeoin` to run the dogecoin
+You can now launch a network with `dfx start --enable-dogeoin` to run the dogecoin
 integration locally.
 
 ### feat: improved the canister snapshot download/upload feature
@@ -20,15 +20,19 @@ Improved the canister snapshot download/upload feature by
 The custom logic was prone to becoming outdated, such as not adapting to changing cycles fees.
 By using `pocket-ic`, which gets updated frequently, the BTC integration is significanly less likely to break.
 
+### fix: `dfx start --enable-bitcoin` will add `--bitcoin-node 127.0.0.1:18444` unless nodes are specified in dfx.json
+
 ### chore: Bump cdk to 0.19 in project template.
 
 ## Dependencies
 
 ### Replica
 
-Updated replica to commit 2e99e6c26c0eacd4f321e7c2604d118d645828ad
+Updated replica to commit 948d5b9260494ec3e6c9bc9db499f34d52ba6c7f
 This incorporates the following executed proposals:
 
+- [139480](https://dashboard.internetcomputer.org/proposal/139480)
+- [139403](https://dashboard.internetcomputer.org/proposal/139403)
 - [139317](https://dashboard.internetcomputer.org/proposal/139317)
 - [139192](https://dashboard.internetcomputer.org/proposal/139192)
 - [139079](https://dashboard.internetcomputer.org/proposal/139079)